pyproject.toml
[tool.poetry]
name = "todolist_server"
version = "0.1.0"
description = ""
authors = ["Sedir Morais <philippi.sedir@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
djangorestframework = "^3.9"
djoser = "^1.7"
django-rest-swagger = "^2.2"
gunicorn = "^19.9"
whitenoise = "^4.1"
dj-database-url = "^0.5.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
pylint = "^2.3"
flake8 = "^3.7"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
